+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## v0.0.x
 
+- v0.0.3:
+  - **Meta:** - Change name to `paper-feeds`
+  - **Frontend:** - [#11](https://github.com/sneakers-the-rat/journal-rss/pull/11) - Improved CSS/HTML structure from [@roaldarbol](https://github.com/roaldarbol) 
+  - **Bugfix** - [#16](https://github.com/sneakers-the-rat/journal-rss/issues/16), [#20](https://github.com/sneakers-the-rat/journal-rss/pull/20) - Filter paper types and properly handle null-like datetimes
 - v0.0.2:
   - Semver begins - all merges to main should increment version
   - Added database migrations with alembic - all model changes should come with migrations after now
-  - **Docs:** Some initial contribution docs. Closes [#1](https://github.com/sneakers-the-rat/journal-rss/pull/1)
+  - **Docs:** Some initial contribution docs. Closes [#1](https://github.com/sneakers-the-rat/paper-feeds/pull/1)
+  - **Frontend:** - [#7](https://github.com/sneakers-the-rat/journal-rss/pull/7) - Fix some early bugs from [@lambdaloop](https://github.com/lambdaloop) 
 - v0.0.1: (prehistory)
