+++ pyproject.toml
@@ -1,16 +1,35 @@
 [tool.poetry]
-name = "journal-rss"
-version = "0.0.2"
-description = "A FastAPI web server for creating RSS feeds for scholarly journals with the magic of adversarial interoperability"
+name = "paper-feeds"
+version = "0.0.3"
+description = "A FastAPI web server for creating RSS and ActivityPub feeds for scholarly literature with the magic of adversarial interoperability"
 authors = ["sneakers-the-rat <JLSaunders987@gmail.com>"]
+repository = "https://github.com/sneakers-the-rat/paper-feeds"
 license = "GPL-3.0"
 readme = "README.md"
 packages = [
-    { include = "journal_rss", from="src" }
+    { include = "paper_feeds", from="src" }
 ]
 include = [
     "src/static/*",
     "src/templates/*"
+]
+keywords = ["rss", "activitypub", "scholarly communication", "scholcomm", "fastapi", "metadata"]
+classifiers = [
+    "Development Status :: 2 - Pre-Alpha",
+    "Framework :: FastAPI",
+    "Framework :: Pydantic :: 2",
+    "Intended Audience :: Science/Research",
+    "Intended Audience :: System Administrators",
+    "Intended Audience :: End Users/Desktop",
+    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+    "Natural Language :: English",
+    "Programming Language :: Python :: 3.11",
+    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: News/Diary",
+    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
+    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
+    "Topic :: Internet :: WWW/HTTP :: WSGI :: Server",
+    "Topic :: Scientific/Engineering",
+    "Topic :: Communications"
 ]
 
 [tool.poetry.dependencies]
@@ -22,7 +41,7 @@ uvicorn = "^0.25.0"
 jinja2 = "^3.1.2"
 requests = "^2.31.0"
 python-multipart = "^0.0.6"
-fastapi-rss = "^0.2.0"
+fastapi-rss = "^0.2.2"
 alembic = "^1.13.1"
 pytest = {version = "^7.4.4", optional = true}
 requests-cache = {version = "^1.1.1", optional = true}
@@ -35,7 +54,7 @@ pytest-timeout = {version = "^2.2.0", optional = true}
 tests = ['pytest', 'requests-cache', 'pytest-cov', 'coveralls', 'pytest-alembic', 'pytest-timeout']
 
 [tool.poetry.scripts]
-start = "journal_rss.main:start"
+start = "paper_feeds.main:start"
 
 [build-system]
 requires = ["poetry-core"]
@@ -43,17 +62,17 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
 addopts = [
-    "--cov=journal_rss",
+    "--cov=paper_feeds",
     "--cov-append",
     "--cov-config=pyproject.toml",
 ]
 
 [tool.coverage.run]
 branch = true
-source = ["src/journal_rss"]
+source = ["src/paper_feeds"]
 omit = [
     "tests/*",
-    "src/journal_rss/static/*",
-    "src/journal_rss/templates/*"
+    "src/paper_feeds/static/*",
+    "src/paper_feeds/templates/*"
 ]
