--- src/journal_rss/__init__.py
@@ -1,4 +1,0 @@
-
-
-from journal_rss.config import Config
-from journal_rss.logging import init_logger
