--- src/journal_rss/__main__.py
@@ -1,3 +1,0 @@
-from journal_rss.main import start
-
-start()
