+++ src/paper_feeds/__main__.py
@@ -1,0 +1,3 @@
+from paper_feeds.main import start
+
+start()
