+++ src/paper_feeds/main.py
@@ -1,11 +1,11 @@
 import uvicorn
 
-from journal_rss.config import Config
+from paper_feeds.config import Config
 
 def start():
     config = Config()
     uvicorn.run(
-        "journal_rss.app:app",
+        "paper_feeds.app:app",
         host=config.host,
         port=config.port,
         reload=config.reload
