+++ src/paper_feeds/models/__init__.py
@@ -3,6 +3,6 @@ All models used in the DB should be imported here so they are created by
 :func:`.db.create_tables`
 """
 from sqlmodel import SQLModel
-from journal_rss.models.journal import Journal, ISSN
-from journal_rss.models.paper import Paper
+from paper_feeds.models.journal import Journal, ISSN
+from paper_feeds.models.paper import Paper
