+++ src/templates/includes/footer.html
@@ -1,5 +1,5 @@
 <footer class="footer_section">
     <p class="source">
-        A <a href="https://github.com/sneakers-the-rat/journal-rss" target="_blank">Journal-RSS</a> instance <3
+        A <a href="https://github.com/sneakers-the-rat/paper-feeds" target="_blank">paper-feeds</a> instance <3
     </p>
 </footer>
